students-CRUD-mockAPIs/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
azure-cosmos==4.5.1
gunicorn==21.2.0
python-dotenv==1.0.0
